Add Transform method to the GfQuat{h,f,d} classes
Transform transforms a point by the quaternion. Given a quaternion q
and a point p,

    q.Transform(p)

it is the equivalent of:

    (q * GfQuat?(0, p) * q.GetInverse()).GetImaginary()

where GfQuat? is the same type as q. The implementation takes advantage
of redundancies in the math to be more efficient than a straightforward
evaluation of the above equation. Note that the more efficient
evaluation is still 22 multiplies, 11 additions, and 1 division. If you
have many points to transform by a single quaternion, it will be more
efficient to construct a matrix from the quaternion and use the matrix
to transform the points.

The test testGfQuaternion used to focus almost all of its attention on
the GfQuaternion class and barely tested the GfQuat? classes. Now
testing of GfQuat? is significantly expanded and tests even more
features than are tested for GfQuaternion.

(Internal change: 2156058)