[gf] testGfMatrix: increase tolerance for float AssertDeterminant test
...and lower it for doubles.  Previously, we used the same tolerance
(1e-6) for both floats and doubles.  This sets the epsilon to 1e-5 for
floats, and 1e-10 for doubles.

Running on an ARM system (Arm Neoverse N1) yielded an
absolute error of 1.0259421223521911e-06 for one of the GetInverse
tests, which exceeded this tolerance.  Testing on a more "standard" AMD
cpu (AMD Ryzen Threadripper PRO 3975WX) yielded an error of
9.610960183499514e-07 for the same test - juuuuust below the threshold.

My guess is that the threshold was never meant to be so tight, and that
1e-6 was chosen arbitrarily, and it seemed to work.  1e-5 should provide
a bit more tolerance for float architecture differences.

At the same time, increasing the tolerance for double matrices gives us
increased safety for the code paths that will generally be used for
matrix math.  (In my tests, the maximum error for both ARM and AMD
architectures never got out of the e-15 range, so e-10 should still be
plenty of headroom.)