Remove GfMatrix4d::TransformAffine overload that potentially loses precision
The overload of this method that takes and returns a GfVec3f can easily cause
loss of precision when the matrix in question represents a local-to-world
transformation. Before this change, client code that transforms a local GfVec3f
to a world GfVec3d had to be careful to first convert the argument to GfVec3f, in
order to select the overload that preserves precision. This change removes the
lossy overload.

This change includes updates to call sites where the result of a call to
TransformAffine now needs to be converted to GfVec3f. The advantage of requiring
these conversions is that the loss of precision is explicit.

(Internal change: 2333663)