Remove GfRotation::TransformDir overload that potentially loses precision
The overload of this methods that takes and returns a GfVec3f can cause
inadvertent loss of precision.

This change includes updates to call sites where the result of a call to
TransformDir now needs to be converted to GfVec3f. The advantage of requiring
these conversions is that the loss of precision is explicit.

(Internal change: 2335200)