Remove matrix multiply operator overloads that potentially lose precision.
The overloads of this methods that take and return a GfVec3f can cause
inadvertent loss of precision.

This change includes updates to call sites where the result of a call to one of
these overloads now needs to be converted to GfVec3f. The advantage of requiring
these conversions is that the loss of precision is explicit.

(Internal change: 2335357)