gf: Temporarily suppress warning C4018 in GfIntegerCompareLess
This code reportedly began emitting this warning in OpenUSD 25.11.
For now, we simply silence this warning rather than make any
low-level changes this late in the release cycle. We expect to remove
this in the subsequent release.

This change also adds the unit test cases for bool conversions in
GfNumericCast from pull request #3855.

See #3855

(Internal change: 2382628)